Cleaned up fragments after merge. Pulled BookListAdapter into it's own class.
