+++ app/src/main/java/com/cs428/app/bookapp/activity/fragments/ProfileFragment.java
@@ -17,6 +17,8 @@ import com.cs428.app.bookapp.model.Book;
 import com.cs428.app.bookapp.model.Model;
 import com.cs428.app.bookapp.model.Person;
 
+import com.cs428.app.bookapp.adapter.BookListAdapter;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
@@ -73,7 +75,7 @@ public class ProfileFragment extends Fragment {
 
         // Give the option to unfollow someone
         else if (person != null && !person.isUser()
-                && Model.SINGLETON.currentUser.isFollowing(person.getId())) {
+                && Model.getSINGLETON().getCurrentUser().isFollowing(person.getId())) {
             actionButton.setText(R.string.unfollow);
         }
 
@@ -85,54 +87,5 @@ public class ProfileFragment extends Fragment {
         super.onDetach();
     }
 
-    public class BookListAdapter extends RecyclerView.Adapter<BookListAdapter.ViewHolder> {
-        private List<Book> books;
 
-        public class ViewHolder extends RecyclerView.ViewHolder {
-            // each data item is just a string in this case
-            public ImageView imageView;
-            public ViewHolder(ImageView v) {
-                super(v);
-                imageView = v;
-            }
-        }
-
-        // Provide a suitable constructor (depends on the kind of dataset)
-        public BookListAdapter(List<Book> books) {
-            this.books = books;
-        }
-
-        // Create new views (invoked by the layout manager)
-        @Override
-        public BookListAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-            // create a new view
-            ImageView v = (ImageView) LayoutInflater.from(parent.getContext())
-                    .inflate(R.layout.book_cover_view, parent, false);
-
-            return new ViewHolder(v);
-        }
-
-        // Replace the contents of a view (invoked by the layout manager)
-        @Override
-        public void onBindViewHolder(ViewHolder holder, int position) {
-            // - get element from dataset at this position
-            // - replace the contents of the view with that element
-
-            //TODO: make sure this works, getCover now returns a url so this was changed to adapt to that
-            String url = books.get(position).getCover();
-            Bitmap bitmap = null;
-            try { bitmap = BitmapFactory.decodeStream((InputStream)new URL(url).getContent());
-            } catch (IOException e) {e.printStackTrace();}
-            holder.imageView.setImageBitmap(bitmap);
-
-            //original code:
-            //holder.imageView.setImageBitmap(books.get(position).getCover());
-        }
-
-        // Return the size of your dataset (invoked by the layout manager)
-        @Override
-        public int getItemCount() {
-            return books.size();
-        }
-    }
 }
