+++ app/src/main/res/layout/fragment_login.xml
@@ -3,8 +3,8 @@
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:padding="10dip"
-        android:layout_below="@id/toolbar">
+        android:padding="10dip">
+        <!--android:layout_below="@id/toolbar">-->
     <!--  Email Label -->
     <TextView android:layout_width="fill_parent"
               android:layout_height="wrap_content"
